rename: 리뷰 페이지 관련 컴포넌트 폴더 정리
